cypress config + seperate text dictionary
