+++ frontend/.babelrc
@@ -1,0 +1,6 @@
+{
+    "presets": [
+      "@babel/preset-env",
+      ["@babel/preset-react", { "runtime": "automatic" }]
+    ]
+  }
