+++ frontend/cypress/e2e/wedding.cy.js
@@ -1,0 +1,8 @@
+import text from '../../src/resources/dictionary';
+
+describe('Site entry', () => {
+  it('Entry page renders', () => {
+    cy.visit('http://localhost:3003');
+    cy.contains(text['enter'].eng);
+  });
+});
